Use spaced endashes `–` instead of hyphens for sentence breaks
Hyphens should be used for compound-words and hyphenations, not sentence breaks.
For sentence breaks either spaced endashes or unspaced emdashes should be used.
In this pull request I implemented endashes, because they look closer to the
past implementation with hyphens. It seems that the spaced endash is a UK style,
while the US uses the emdashes for sentence breaks (see second link below).

I am not an experienced typhographer, but this is my experience from editing
Wikipedia, see
https://en.wikipedia.org/wiki/Wikipedia:Manual_of_Style/Dash_draft.
But a google search showed consistent recommendations in other style
recommendations such as
https://www.scribbr.com/language-rules/dashes/

I only looked at the new FSRS tooltips, didn't check all strings for wrong
hyphens. Just also fixed that in the readme because I noticed it by accident
there. Though honestly I don't care for hyphen-lengths in technical
documentation all that much, just for proper typography in user-facing texts.